реализовано все, кроме модальных окон
