+++ src/components/burger-constructor/burger-constructor.module.css
@@ -2,7 +2,11 @@
     display: flex;
     flex-direction: column;
     flex: 1 1 auto;
+    flex-grow: 1;
+    width: 600px;
+    align-items:stretch;
 }
+
 .scrollable{
     flex:1 1 auto;
     margin: 1 auto;
@@ -12,7 +16,7 @@
     }
 .scrollable::-webkit-scrollbar {
     width: 8px;
-    background-color: #f9f9fd;
+    background-color: #2F2F37;
   }
 .scrollable::-webkit-scrollbar-thumb {
     background-color: #8585ad;
