+++ src/components/burger-element/burger-element.module.css
@@ -1,0 +1,7 @@
+.element{
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    flex-shrink: 1;
+
+}
