+++ src/components/burger-ingredient-card/burger-ingredient-card.module.css
@@ -1,0 +1,20 @@
+.ingredient_card{
+    display: flex;
+    flex-direction:column;
+    width: 280px;
+    position: relative;
+}
+.price_div{
+    display: flex;
+    flex-direction: row;
+    flex: 1 1 auto;
+    margin: 0 auto;
+  }
+.item_center{
+   flex: 1 1 auto;
+   text-align: center;
+}
+.img{
+  width: 240px;
+  height: 130px;
+}
