+++ src/components/burger-ingredients-tab/burger-ingredients-tab.module.css
@@ -1,0 +1,4 @@
+.ingretients_tab{
+    display:flex;
+    flex-wrap:wrap;
+}
