+++ src/components/modal/modal.js
@@ -1,21 +1,22 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-const modalRoot = document.getElementById("react-modals");
 
 class Modal extends React.Component {
   render() {
     const { children, header, onClose } = this.props;
+    const modalRoot = document.getElementById("react-modals");
 
-    return ReactDOM.createPortal(
+    return (ReactDOM.createPortal(
       <>
-        <div style={{position:'absolute', margin:'0 auto'}}>
+        <div style={{position:'absolute', margin:'0 auto', border:'2px'}}>
           <div onClose={onClose}>{header}</div>
           {children}
         </div>
         <span onClose={onClose} />
       </>,
       modalRoot
-    );
+    )
+    )
   }
 } 
 export default Modal
