+++ src/components/submit-order/submit-order.module.css
@@ -1,0 +1,7 @@
+.submit{
+    display: flex;
+    flex-direction: row;    
+    justify-content: flex-end;
+    align-items: center;
+    align-self: center;
+}
